refactor(entry): split agent server facade into submodules
